requirements.txt
asgiref==3.2.9
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
Deprecated==1.2.10
Django==3.0.7
django-allauth==0.42.0
idna==2.9
oauthlib==3.1.0
PyGithub==1.51
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
urllib3==1.25.9
wrapt==1.12.1